feat: container names from project path
If name is omitted from .dev-env/{environment}/dev-env.json then use a
name based on the project path. This allows .dev-env directories to be
copied between projects without the new project re-using the container
of the original project. If an explicit name is provided that name will
be used preserving the original usage if desired.